+++ src/assets/scripts/client/aircraft/AircraftInstanceModel.js
@@ -1,20 +1,27 @@
 /* eslint-disable max-len, no-undef */
-import $ from 'jquery';
-import _clamp from 'lodash/clamp';
 import _defaultTo from 'lodash/defaultTo';
 import _forEach from 'lodash/forEach';
 import _get from 'lodash/get';
 import _isEqual from 'lodash/isEqual';
 import _isNil from 'lodash/isNil';
+import _random from 'lodash/random';
 import _uniqueId from 'lodash/uniqueId';
-import AircraftStripView from './AircraftStripView';
 import Fms from './FlightManagementSystem/Fms';
 import ModeController from './ModeControl/ModeController';
 import Pilot from './Pilot/Pilot';
 import { speech_say } from '../speech';
-import { tau, radians_normalize, angle_offset } from '../math/circle';
-import { abs, cos, extrapolate_range_clamp, sin, spread } from '../math/core';
-import { getOffset, calculateTurnInitiaionDistance } from '../math/flightMath';
+import { radians_normalize, angle_offset } from '../math/circle';
+import {
+    getOffset,
+    calculateTurnInitiaionDistance
+} from '../math/flightMath';
+import {
+    abs,
+    cos,
+    extrapolate_range_clamp,
+    sin,
+    spread
+} from '../math/core';
 import {
     distance_to_poly,
     point_to_mpoly,
@@ -34,19 +41,17 @@ import {
 } from '../utilities/radioUtilities';
 import {
     degreesToRadians,
-    heading_to_string,
-    km,
-    nm,
-    UNIT_CONVERSION_CONSTANTS
+    nm
 } from '../utilities/unitConverters';
 import {
     FLIGHT_CATEGORY,
     FLIGHT_PHASE,
-    PERFORMANCE,
-    WAYPOINT_NAV_MODE
+    PERFORMANCE
 } from '../constants/aircraftConstants';
-import { AIRPORT_CONSTANTS, AIRPORT_CONTROL_POSITION_NAME } from '../constants/airportConstants';
-import { SELECTORS } from '../constants/selectors';
+import {
+    AIRPORT_CONSTANTS,
+    AIRPORT_CONTROL_POSITION_NAME
+} from '../constants/airportConstants';
 import { GAME_EVENTS } from '../game/GameController';
 import { MCP_MODE, MCP_MODE_NAME } from './ModeControl/modeControlConstants';
 import { TIME } from '../constants/globalConstants';
@@ -75,14 +80,15 @@ export default class AircraftInstanceModel {
      */
     constructor(options = {}, navigationLibrary) {
         /* eslint-disable no-multi-spaces*/
-        this._id = _uniqueId('aircraft-');
+        this.id = _uniqueId('aircraft-');
         this._navigationLibrary = navigationLibrary;
         this.positionModel = null;       // Aircraft Position
         this.model        = null;       // Aircraft type
         this.airlineId      = '';         // Airline Identifier (eg. 'AAL')
         this.airlineCallsign = '';
         this.flightNumber = '';         // Flight Number ONLY (eg. '551')
-        this.transponderCode = '';
+        // FIXME: temporary inclusion, should be removed before merging to develop
+        this.transponderCode = _random(1201, 5999);
         this.heading      = 0;          // Magnetic Heading
         this.altitude     = 0;          // Altitude, ft MSL
         this.speed        = 0;          // Indicated Airspeed (IAS), knots
@@ -93,14 +99,35 @@ export default class AircraftInstanceModel {
         this.approachDistance = 0;      // Distance longitudinally from the threshold
         this.radial       = 0;          // Angle from airport center to aircraft
         this.distance     = 0;          //
-        this.destination  = null;       // Destination they're flying to
+
+        /**
+         * The origin ariport for an aircraft
+         *
+         * This will only be populated for dpearture aircraft
+         *
+         * @property origin
+         * @type {string}
+         * @default ''
+         */
+        this.origin = '';
+
+        /**
+         * The destination airpot of an aircraft
+         *
+         * This will only be populated for arrivals
+         *
+         * @property destination
+         * @type {string}
+         * @default ''
+         */
+        this.destination = '';
+
         this.trend        = 0;          // Indicator of descent/level/climb (1, 0, or 1)
         this.history      = [];         // Array of previous positions
         this.restricted   = { list: [] };
         this.notice       = false;      // Whether aircraft
         this.warning      = false;      //
         this.hit          = false;      // Whether aircraft has crashed
-        this.taxi_next    = false;      //
         this.taxi_start   = 0;          //
         this.taxi_time    = 3;          // Time spent taxiing to the runway. *NOTE* this should be INCREASED to around 60 once the taxi vs LUAW issue is resolved (#406)
         this.rules        = FLIGHT_RULES.IFR;      // Either IFR or VFR (Instrument/Visual Flight Rules)
@@ -136,10 +163,6 @@ export default class AircraftInstanceModel {
          */
         this.isRemovable = false;
 
-        // TODO: change name, and update refs in `InputController`. perhaps change to be a ref to the AircraftStripView class instead of directly accessing the html?
-        this.aircraftStripView = null;
-        this.$html = null;
-        this.$strips = $(SELECTORS.DOM_SELECTORS.STRIPS);
         /* eslint-enable multi-spaces*/
 
         // Set to true when simulating future movements of the aircraft
@@ -175,7 +198,6 @@ export default class AircraftInstanceModel {
             ? window.gameController.game_time()
             : null;
 
-
         this.buildCurrentTerrainRanges();
         this.buildRestrictedAreaLinks();
         this.parse(options);
@@ -194,9 +216,6 @@ export default class AircraftInstanceModel {
 
             this.mcp.initializeForAirborneFlight(bottomAltitude, this.heading, this.speed);
         }
-
-        this.createStrip();
-        this.updateStrip();
     }
 
     /**
@@ -274,7 +293,7 @@ export default class AircraftInstanceModel {
     }
 
     parse(data) {
-        // TODO: these _gets can likely be removed
+        // TODO: the `_get()` below are doing the same thing as a `_defaultTo()` would do
         this.positionModel = _get(data, 'positionModel', this.positionModel);
         this.model = _get(data, 'model', this.model);
         this.airlineId = _get(data, 'airline', this.airlineId);
@@ -284,18 +303,19 @@ export default class AircraftInstanceModel {
         this.heading = _get(data, 'heading', this.heading);
         this.altitude = _get(data, 'altitude', this.altitude);
         this.speed = _get(data, 'speed', this.speed);
+        this.origin = _get(data, 'origin', this.origin);
         this.destination = _get(data, 'destination', this.destination);
+
+        // TODO: this assumes and arrival spawns outside the airspace
         this.inside_ctr = data.category === FLIGHT_CATEGORY.DEPARTURE;
     }
 
     initFms(data) {
         const airport = window.airportController.airport_get();
         const initialRunway = airport.getActiveRunwayForCategory(this.category);
-
         this.fms = new Fms(data, initialRunway, this.model, this._navigationLibrary);
 
         if (this.category === FLIGHT_CATEGORY.DEPARTURE) {
-
             this.setFlightPhase(FLIGHT_PHASE.APRON);
             this.altitude = airport.positionModel.elevation;
             this.speed = 0;
@@ -308,6 +328,45 @@ export default class AircraftInstanceModel {
         if (data.nextFix) {
             this.fms.skipToWaypoint(data.nextFix);
         }
+    }
+
+    /**
+     * Build an object that contains all the correct data, in the correct shape,
+     * so it can be injected into the view.
+     *
+     * This method should only be used by the `StripView` classes when instantiating
+     * or updating the aircraft progress strips.
+     *
+     * The data here should be considered read-only.
+     *
+     * @for AircraftModel
+     * @method getViewModel
+     * @return {object<string, string>}
+     */
+    getViewModel() {
+        let assignedAltitude = this.mcp.altitude;
+        let flightPlanAltitude = this.fms.flightPlanAltitude;
+
+        if (assignedAltitude === -1) {
+            assignedAltitude = '-';
+        }
+
+        if (flightPlanAltitude === -1) {
+            flightPlanAltitude = '-';
+        }
+
+        return {
+            id: this.id,
+            insideCenter: this.inside_ctr,
+            callsign: this.callsign,
+            transponderCode: this.transponderCode,
+            icaoWithWeightClass: this.model.icaoWithWeightClass,
+            assignedAltitude,
+            flightPlanAltitude,
+            arrivalAirportId: this.destination.toUpperCase(),
+            departureAirportId: this.origin.toUpperCase(),
+            flightPlan: this.fms.getFlightPlanRouteForStripView()
+        };
     }
 
     /**
@@ -326,7 +385,6 @@ export default class AircraftInstanceModel {
 
         // Crossing into the center
         if (this.inside_ctr) {
-            this.showStrip();
             this.callUp();
 
             return;
@@ -345,13 +403,11 @@ export default class AircraftInstanceModel {
             return this.arrivalExit();
         }
 
-        this.hideStrip();
         this.setIsRemovable();
 
-        // TODO: this seems redundant. if its already in the leg its in the fms.
-        if (this.mcp.headingMode !== MCP_MODE.HEADING.LNAV || !this.fms.hasWaypoint(this.fms.currentLeg.exitName)) {
+        if (this.mcp.headingMode !== MCP_MODE.HEADING.LNAV) {
             this.radioCall(
-                `leaving radar coverage without being cleared to ${this.fms.currentLeg.exitName}`,
+                'leaving radar coverage without proper clearance',
                 AIRPORT_CONTROL_POSITION_NAME.DEPARTURE,
                 true
             );
@@ -517,6 +573,14 @@ export default class AircraftInstanceModel {
         return !this.isOnGround();
     }
 
+    isDeparture() {
+        return this.fms.isDeparture();
+    }
+
+    isArrival() {
+        return this.fms.isArrival();
+    }
+
     /**
      * Aircraft is established on the course tuned into the nav radio and course buildCurrentTerrainRanges
      *
@@ -525,21 +589,17 @@ export default class AircraftInstanceModel {
      * @return {boolean}
      */
     isEstablishedOnCourse() {
-        const runway = this.fms.arrivalRunway;
+        const runwayModel = this.fms.arrivalRunwayModel;
 
-        if (!runway) {
+        if (!runwayModel) {
             return false;
         }
 
-        const runwayHeading = runway.angle;
-        const approachOffset = getOffset(this, runway.relativePosition, runwayHeading);
-        const lateralDistanceFromCourse_nm = abs(nm(approachOffset[0]));
-        const onApproachCourse = lateralDistanceFromCourse_nm <= PERFORMANCE.MAXIMUM_DISTANCE_CONSIDERED_ESTABLISHED_ON_APPROACH_COURSE_NM;
-        const heading_diff = abs(angle_offset(this.heading, runwayHeading));
-        const onCorrectHeading = heading_diff < PERFORMANCE.MAXIMUM_ANGLE_CONSIDERED_ESTABLISHED_ON_APPROACH_COURSE;
+        // TODO: the `this` here is ugly, but will be needed until `getOffset`
+        // is refactored (#291 - https://github.com/openscope/openscope/issues/291)
+        return runwayModel.isOnApproachCourse(this) && runwayModel.isOnCorrectApproachHeading(this.heading);
 
-        return onApproachCourse && onCorrectHeading;
-
+        // TODO: Use this instead
         // const courseDatum = this.mcp.nav1Datum;
         // const course = this.mcp.course;
         // const courseOffset = getOffset(this, courseDatum.relativePosition, course);
@@ -580,8 +640,7 @@ export default class AircraftInstanceModel {
     isOnGround() {
         const errorAllowanceInFeet = 5;
         const airport = window.airportController.airport_get();
-        const runway = this.fms.currentRunway;
-        const nearRunwayAltitude = abs(this.altitude - runway.elevation) < errorAllowanceInFeet;
+        const nearRunwayAltitude = abs(this.altitude - this.fms.currentRunway.elevation) < errorAllowanceInFeet;
         const nearAirportAltitude = abs(this.altitude - airport.elevation) < errorAllowanceInFeet;
 
         return nearRunwayAltitude || nearAirportAltitude;
@@ -629,24 +688,20 @@ export default class AircraftInstanceModel {
         return this.isTaxiing() || this.flightPhase === FLIGHT_PHASE.TAKEOFF;
     }
 
-    // TODO: the logic in this method can be cleaned up and simplified
     /**
      * @for AircraftInstanceModel
      * @method isVisible
+     * @return {boolean}
      */
     isVisible() {
-        // TODO: this if/else if would be cleaner with just if (this.flightPhase === FLIGHT_PHASE.WAITING) {}
-        // hide aircraft on twys
+        // hide aircraft on taxiways
         if (this.flightPhase === FLIGHT_PHASE.APRON || this.flightPhase === FLIGHT_PHASE.TAXI) {
             return false;
         }
 
-        if (this.isTaxiing()) {
+        if (this.flightPhase === FLIGHT_PHASE.WAITING) {
             // show only the first aircraft in the takeoff queue
-            const runway = this.fms.departureRunway;
-            const nextInRunwayQueue = runway.isAircraftNextInQueue(this);
-
-            return this.flightPhase === FLIGHT_PHASE.WAITING && nextInRunwayQueue;
+            return this.fms.departureRunwayModel.isAircraftNextInQueue(this.id);
         }
 
         return true;
@@ -655,7 +710,9 @@ export default class AircraftInstanceModel {
     /**
      * Sets `#isRemovable` to true
      *
-     * Provides a single source to change the value of `#isRemovable`.
+     * Provides a single source to change the value of `#isRemovable`
+     * This is evaluated by the `AircraftController` when determining
+     * if an aircraft should be removed or not
      *
      * @for AircraftInstanceModel
      * @method setIsRemovable
@@ -675,10 +732,9 @@ export default class AircraftInstanceModel {
             head: 0
         };
 
-        const airport = window.airportController.airport_get();
-        const wind = airport.wind;
-        const runway = this.fms.currentRunway;
-        const angle =  abs(angle_offset(runway.angle, wind.angle));
+        const { wind } = window.airportController.airport_get();
+        // const wind = airport.wind;
+        const angle = this.fms.currentRunway.calculateCrosswindAngleForRunway(wind.angle);
 
         // TODO: these two bits of math should be abstracted to helper functions
         windForRunway.cross = sin(angle) * wind.speed;
@@ -695,7 +751,8 @@ export default class AircraftInstanceModel {
      * @param runwayModel {RunwayModel}
      */
     moveToRunway(runwayModel) {
-        this.positionModel.setCoordinates(runwayModel.positionModel.gps);
+        this.positionModel.setCoordinates(runwayModel.gps);
+
         this.heading = runwayModel.angle;
         this.altitude = runwayModel.elevation;
     }
@@ -954,8 +1011,7 @@ export default class AircraftInstanceModel {
      * @method updateFlightPhase
      */
     updateFlightPhase() {
-        const airportModel = window.airportController.airport_get();
-        const runwayModel = this.fms.departureRunway;
+        const runwayModel = this.fms.departureRunwayModel;
 
         if (this._shouldEnterHoldingPattern()) {
             this.setFlightPhase(FLIGHT_PHASE.HOLD);
@@ -1214,10 +1270,10 @@ export default class AircraftInstanceModel {
         // const altitudeToTarget = _clamp(glideslopeAltitude, glideDatum.elevation, this.altitude);
 
         // ILS SPECIFIC CODE
-        const runway = this.fms.arrivalRunway;
-        const offset = getOffset(this, runway.relativePosition, runway.angle);
-        const distanceOnFinal_km = offset[1];
-        const glideslopeAltitude = runway.getGlideslopeAltitude(distanceOnFinal_km);
+        const runwayModel = this.fms.arrivalRunwayModel;
+        const offset = getOffset(this, runwayModel.relativePosition, runwayModel.angle);
+        const distanceOnFinalKm = offset[1];
+        const glideslopeAltitude = runwayModel.getGlideslopeAltitude(distanceOnFinalKm);
         const altitudeToTarget = Math.min(this.mcp.altitude, glideslopeAltitude);
 
         return altitudeToTarget;
@@ -1235,18 +1291,20 @@ export default class AircraftInstanceModel {
         const datum = this.mcp.nav1Datum;
         const course = this.mcp.course;
         const courseOffset = getOffset(this, datum.relativePosition, course);
-        const lateralDistanceFromCourse_nm = nm(courseOffset[0]);
+        const lateralDistanceFromCourseNm = nm(courseOffset[0]);
         const headingDifference = angle_offset(course, this.heading);
         const bearingFromAircaftToRunway = this.positionModel.bearingToPosition(datum);
         const angleAwayFromLocalizer = course - bearingFromAircaftToRunway;
         const turnTimeInSeconds = abs(headingDifference) / PERFORMANCE.TURN_RATE;    // time to turn headingDifference degrees
+        // TODO: this should be moved to a class method `.getTurningRadius()`
         const turningRadius = this.speed * (turnTimeInSeconds * TIME.ONE_SECOND_IN_HOURS);  // dist covered in the turn, nm
         const distanceCoveredDuringTurn = turningRadius * abs(headingDifference);
-        const distanceToLocalizer = lateralDistanceFromCourse_nm / sin(headingDifference); // dist from the localizer intercept point, nm
+        const distanceToLocalizer = lateralDistanceFromCourseNm / sin(headingDifference); // dist from the localizer intercept point, nm
         const distanceEarly = 0.5;    // start turn early, to avoid overshoots from tailwind
         const shouldAttemptIntercept = (distanceToLocalizer > 0 && distanceToLocalizer <= distanceCoveredDuringTurn + distanceEarly);
         const inTheWindow = abs(angleAwayFromLocalizer) < degreesToRadians(1.5);  // if true, aircraft will move to localizer, regardless of assigned heading
 
+        // TODO: this logic is confusing, simplify
         if (!(shouldAttemptIntercept || inTheWindow)) {
             return this.mcp.heading;
         }
@@ -1254,7 +1312,7 @@ export default class AircraftInstanceModel {
         const severity_of_correction = 50;  // controls steepness of heading adjustments during localizer tracking
         let interceptAngle = angleAwayFromLocalizer * -severity_of_correction;
         const minimumInterceptAngle = degreesToRadians(10);
-        const isAlignedWithCourse = abs(lateralDistanceFromCourse_nm) <= PERFORMANCE.MAXIMUM_DISTANCE_CONSIDERED_ESTABLISHED_ON_APPROACH_COURSE_NM;
+        const isAlignedWithCourse = abs(lateralDistanceFromCourseNm) <= PERFORMANCE.MAXIMUM_DISTANCE_CONSIDERED_ESTABLISHED_ON_APPROACH_COURSE_NM;
 
         // TODO: This is a patch fix, and it stinks. This whole method needs to be improved greatly.
         if (isAlignedWithCourse) {
@@ -1287,7 +1345,6 @@ export default class AircraftInstanceModel {
     updateLandingFailedLanding() {
         // Failed Approach
         if ((this.approachDistance > 0.100) && (!this.projected)) {
-            this.updateStrip();
             this.cancelLanding();
 
             const isWarning = true;
@@ -1418,15 +1475,15 @@ export default class AircraftInstanceModel {
      * @return {number}
      */
     _calculateTargetedAltitudeDuringLanding() {
-        const runway = this.fms.arrivalRunway;
-        const offset = getOffset(this, runway.relativePosition, runway.angle);
+        const runwayModel = this.fms.arrivalRunwayModel;
+        const offset = getOffset(this, runwayModel.relativePosition, runwayModel.angle);
         const distanceOnFinal_km = offset[1];
 
         if (distanceOnFinal_km > 0) {
             return this._calculateTargetedAltitudeToInterceptGlidepath();
         }
 
-        return runway.elevation;
+        return runwayModel.elevation;
     }
 
     /**
@@ -1437,17 +1494,17 @@ export default class AircraftInstanceModel {
      * @return {number}
      */
     _calculateTargetedHeadingDuringLanding() {
-        const runway = this.fms.arrivalRunway;
-        const offset = getOffset(this, runway.relativePosition, runway.angle);
+        const runwayModel = this.fms.arrivalRunwayModel;
+        const offset = getOffset(this, runwayModel.relativePosition, runwayModel.angle);
         const distanceOnFinal_nm = nm(offset[1]);
 
         if (distanceOnFinal_nm > 0) {
-            const bearingFromAircaftToRunway = this.positionModel.bearingToPosition(runway.positionModel);
+            const bearingFromAircaftToRunway = this.positionModel.bearingToPosition(runwayModel.positionModel);
 
             return bearingFromAircaftToRunway;
         }
 
-        return runway.angle;
+        return runwayModel.angle;
     }
 
     /**
@@ -1459,8 +1516,8 @@ export default class AircraftInstanceModel {
      */
     _calculateTargetedSpeedDuringLanding() {
         let startSpeed = this.speed;
-        const runway  = this.fms.arrivalRunway;
-        const offset = getOffset(this, runway.relativePosition, runway.angle);
+        const runwayModel = this.fms.arrivalRunwayModel;
+        const offset = getOffset(this, runwayModel.relativePosition, runwayModel.angle);
         const distanceOnFinal_nm = nm(offset[1]);
 
         if (distanceOnFinal_nm <= 0 && this.isOnGround())  {
@@ -1828,8 +1885,9 @@ export default class AircraftInstanceModel {
                 }
             }
 
+            // this was a $.each() and may need to verified that its working with _forEach()
             // raise warning if in at least one restricted area
-            $.each(this.restricted.list, (k, v) => {
+            _forEach(this.restricted.list, (k, v) => {
                 warning = warning || v.inside;
             });
         }
@@ -1921,7 +1979,6 @@ export default class AircraftInstanceModel {
         this.updateFlightPhase();
         this.updateTarget();
         this.updatePhysics();
-        this.updateStrip();
     }
 
     /**
@@ -1973,152 +2030,5 @@ export default class AircraftInstanceModel {
      */
     removeConflict(conflictingAircraft) {
         delete this.conflicts[conflictingAircraft.callsign];
-    }
-
-    // TODO: aircraft strip methods below will be abstracted and de-coupled from this model
-
-    /**
-     * Create the aircraft's flight strip and add to strip bay
-     */
-    createStrip() {
-        this.aircraftStripView = new AircraftStripView(this);
-
-        this.$html = this.aircraftStripView.$element;
-        // Add the strip to the html
-        const scrollPos = this.$strips.scrollTop();
-        this.$strips.prepend(this.aircraftStripView.$element);
-        // shift scroll down one strip's height
-        this.$strips.scrollTop(scrollPos + this.aircraftStripView.height);
-
-        // Determine whether or not to show the strip in our bay
-        if (this.category === FLIGHT_CATEGORY.ARRIVAL) {
-            this.aircraftStripView.hide();
-        }
-    }
-
-    // TODO: move these view methods to `AircraftStripView` or a different file
-    /**
-     * @for AircraftInstanceModel
-     * @method updateStrip
-     */
-    updateStrip() {
-        if (this.projected || !this.inside_ctr) {
-            return;
-        }
-
-        const heading = heading_to_string(this.mcp.heading);
-        const altitude = this.mcp.altitude;
-        const speed = this.mcp.speed;
-        let destinationDisplay = this.fms.getProcedureAndExitName();
-        const altitudeText = this.taxi_next
-            ? 'ready'
-            : null;
-        const hasAltitude = this.mcp.altitude !== -1;
-
-        this.aircraftStripView.update(heading, altitude, this.destination, speed);
-
-        switch (this.flightPhase) {
-            case FLIGHT_PHASE.APRON:
-                this.aircraftStripView.updateViewForApron(destinationDisplay, hasAltitude);
-
-                break;
-            case FLIGHT_PHASE.TAXI:
-                this.aircraftStripView.updateViewForTaxi(destinationDisplay, hasAltitude, altitudeText);
-
-                break;
-            case FLIGHT_PHASE.WAITING:
-                this.aircraftStripView.updateViewForWaiting(
-                    destinationDisplay,
-                    this.pilot.hasDepartureClearance,
-                    hasAltitude
-                );
-
-                break;
-            case FLIGHT_PHASE.TAKEOFF:
-                this.aircraftStripView.updateViewForTakeoff(destinationDisplay);
-
-                break;
-            case FLIGHT_PHASE.CLIMB:
-            case FLIGHT_PHASE.HOLD:
-            case FLIGHT_PHASE.DESCENT:
-            case FLIGHT_PHASE.CRUISE:
-                let cruiseNavMode = WAYPOINT_NAV_MODE.FIX;
-                let headingDisplay = this.fms.currentWaypoint.name.toUpperCase();
-                const isFollowingSid = this.fms.isFollowingSid();
-                const isFollowingStar = this.fms.isFollowingStar();
-                const fixRestrictions = {
-                    altitude: this.fms.currentWaypoint.altitudeRestriction !== -1,
-                    speed: this.fms.currentWaypoint.speedRestriction !== -1
-                };
-
-                // TODO: this will need to be addressed when the AircraftStripView is refactored
-                // this block is a bandaid to prevent `destinationDisplay` from being undefined
-                destinationDisplay = this.fms.currentWaypoint.name;
-
-                if (isFollowingStar || isFollowingSid) {
-                    destinationDisplay = this.fms.getProcedureAndExitName();
-                }
-
-                if (this.fms.currentWaypoint.isHold) {
-                    cruiseNavMode = WAYPOINT_NAV_MODE.HOLD;
-                    headingDisplay = 'holding';
-                    destinationDisplay = this.fms.getDestinationName();
-                } else if (this.mcp.headingMode === MCP_MODE.HEADING.HOLD) {
-                    headingDisplay = this.mcp.headingInDegrees;
-                    destinationDisplay = this.fms.getDestinationName();
-                } else if (this.mcp.headingMode === MCP_MODE.HEADING.VOR_LOC) {
-                    cruiseNavMode = WAYPOINT_NAV_MODE.RWY;
-                    headingDisplay = 'intercept';
-                    destinationDisplay = this.fms.getDestinationAndRunwayName();
-                }
-
-                this.aircraftStripView.updateViewForCruise(
-                    cruiseNavMode,
-                    headingDisplay,
-                    destinationDisplay,
-                    isFollowingSid,
-                    isFollowingStar,
-                    fixRestrictions
-                );
-
-                break;
-            case FLIGHT_PHASE.APPROACH:
-            case FLIGHT_PHASE.LANDING:
-                destinationDisplay = this.fms.getDestinationAndRunwayName();
-
-                this.aircraftStripView.updateViewForLanding(destinationDisplay);
-
-                break;
-            default:
-                throw new TypeError(`Invalid FLIGHT_MODE ${this.flightPhase} passed to .updateStrip()`);
-        }
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method showStrip
-     */
-    showStrip() {
-        this.$html.detach();
-
-        const scrollPos = this.$strips.scrollTop();
-
-        this.$strips.prepend(this.$html);
-        this.$html.show();
-        // TODO enumerate the magic number
-        // shift scroll down one strip's height
-        this.$strips.scrollTop(scrollPos + 45);
-    }
-
-    /**
-     * @for AircraftInstanceModel
-     * @method hideStrip
-     */
-    hideStrip() {
-        this.$html.hide(600);
-    }
-
-    cleanup() {
-        this.$html.remove();
     }
 }