+++ src/assets/scripts/client/game/GameController.js
@@ -387,4 +387,19 @@ export default class GameController {
     complete() {
         this.game.paused = false;
     }
+
+    /**
+     * Facade for `game.option.get`
+     *
+     * Allows for classes that import the `GameController` single-level
+     * access to any game option value
+     *
+     * @for GameController
+     * @method getGameOption
+     * @param optionName {string}
+     * @return {string}
+     */
+    getGameOption(optionName) {
+        return this.game.option.get(optionName);
+    }
 }