+++ CHANGELOG.md
@@ -2,13 +2,22 @@
 ---
 ### Features
 - Add capability for vectors in route strings [#310](https://github.com/openscope/openscope/issues/310)
-- Adds more contex to the Model classes by adding an optional input paramater [#138](https://github.com/openscope/openscope/issues/138)
+- Adds more context to the Model classes by adding an optional input paramater [#138](https://github.com/openscope/openscope/issues/138)
 - Adds object helper class for object validation  [#191](https://github.com/openscope/openscope/issues/191)
 - Add capability for fly-over fixes in route strings [#19](https://github.com/openscope/openscope/issues/19)
 - Adds squawk/sq command [#372](https://github.com/openscope/openscope/issues/372)
 - Adds the ability to call an airplane by its callsign [#40](https://github.com/openscope/openscope/issues/40)
 - Adds `EventBus` and `EventModel` [#457](https://github.com/openscope/openscope/issues/457)
-
+- Adds `RunwayCollection` and `RunwayRelationshipModel` and moves some runway logic to live in these new classes [#93](https://github.com/openscope/openscope/issues/93)
+    - Abstracts headwind/crosswind calculations to RunwayModel [#312](https://github.com/openscope/openscope/issues/312)
+    - Removes circular reference in AirportModel.runway.airportModel [#58](https://github.com/openscope/openscope/issues/58)
+- Updates `SpawnPatternModel` to use the `AirportModel.arrivalRunway` property when gathering waypoint models needed to calculate initial aircraft heading [#469](https://github.com/openscope/openscope/issues/469)
+- Adds support for suffixes in SID and STAR procedures [#33](https://github.com/openscope/openscope/issues/33)
+- Adds game option to include/hide WIP airports in the airport list [#476](https://github.com/openscope/openscope/issues/476)
+- Adds `StripViewController`, `StripViewCollection`, and `StripViewModel` classes [#285](https://github.com/openscope/openscope/issues/285)
+    - Removes progress strip logic from the `AircraftModel`
+    - Completely reworks CSS for `StripViewList`
+- Updates `AircraftModel.onAirspaceExit()` to look only at the `mcp.headingMode` value [#477](https://github.com/openscope/openscope/issues/477)
 
 
 
@@ -18,7 +27,10 @@
 ### Bugfixes
 - Fixes coordinate letter issue at SBGL [#385](https://github.com/openscope/openscope/issues/385)
 - Prevent NaNs being passed on if invalid altitude is given [#424](https://github.com/openscope/openscope/issues/424)
+- Removes fix command from tutorial and replaces it with infomation on 'route', 'Say Route', and 'Proceed Direct' [#356](https://github.com/openscope/openscope/issues/356)
 - Fixes coordinate letter issues at RJBB, OSDI, OTHH [#325](https://github.com/openscope/openscope/issues/325)
+- Removes KBOS fixes from EKCH [#448](https://github.com/openscope/openscope/issues/448)
+- Runway, wind and spawnPattern changes to allow EGNM to operate [#492](https://github.com/openscope/openscope/issues/492)